Spread out components, task adding functional
