+++ src/style.css
@@ -127,7 +127,7 @@ body{
     border-style: solid;
     border-width: 3px;
     font-family: Georgia;
-    font-size: 40px;
+    font-size: 30px;
     opacity: 62%;
     padding: 5px;
 }
@@ -156,7 +156,7 @@ body{
     border-color:bisque;
     font-family: Georgia;
     font-weight: bold;
-
+    height:100%;
 }
 
 #deleteCompletedTasks{
